Added M1028
Added the high level version of the shotgun. Should be fine for today. It can interact with the current cards. However along with birdshot, a defensive card and dragonsbreath a flame card ( i am not sure how to implement burn damage over turns or how to get a shot to cancel out a type of attack) I am also unable to get the High level shotgun to do extra damage as intended.