Added Incomplete Blocker
Added all three enemies, however, they are not fully functioning. The rocket guy is the only one that is complete since they are a high level sniper replacement that dose more damage. However, the scavanging functionality of the Scavva and the Blocking Functionality of the Blocker has not been implemented yet.